move scons-configure-cache logic into SCons.Utilities.ConfigureCache:main and update references and add entry_point in setup.cfg
