+++ setup.cfg
@@ -54,6 +54,7 @@ packages = find:
 console_scripts =
     scons = SCons.Script.Main:main
     sconsign = SCons.Utilities.sconsign:main
+    scons-configure-cache = SCons.Utilities.ConfigureCache:main
 
 
 [options.package_data]
